requirements.txt
certifi==2021.10.8
charset-normalizer==2.0.7
click==8.0.3
colorama==0.4.4
Flask==2.0.2
fpdf==1.7.2
idna==3.3
itsdangerous==2.0.1
Jinja2==3.0.2
MarkupSafe==2.0.1
Pillow==8.4.0
pywin32==302
requests==2.26.0
urllib3==1.26.7
Werkzeug==2.0.2
win32printing==0.1.3
